# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Wrenn API
  description: AI agent execution platform API.
  version: "0.2.0"
servers:
  - url: http://localhost:8080
    description: Local development
security: []
paths:
...
  /v1/snapshots:
    post:
      summary: Create a snapshot template
      operationId: createSnapshot
      tags: [snapshots]
      security:
        - apiKeyAuth: []
        - sessionAuth: []
      description: |
        Live snapshot: briefly pauses the capsule, writes its VM state +
        memory + flattened rootfs to a new template directory, then resumes
        the capsule. The source capsule keeps running after the snapshot;
        the resulting template can be used to create new capsules.

        Snapshots are immutable: each call must use a fresh name. Re-using
        an existing name returns 409 Conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSnapshotRequest"
      responses:
        "201":
          description: Snapshot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "409":
          description: Name already exists or capsule not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List templates for your team
      operationId: listSnapshots
      tags: [snapshots]
      security:
        - apiKeyAuth: []
        - sessionAuth: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [base, snapshot]
          description: Filter by template type.
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
  /v1/snapshots/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string

    delete:
      summary: Delete a snapshot template
      operationId: deleteSnapshot
      tags: [snapshots]
      security:
        - apiKeyAuth: []
        - sessionAuth: []
      description: Removes the snapshot files from disk and deletes the database record.
      responses:
        "204":
          description: Snapshot deleted
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for capsule lifecycle operations. Create via POST /v1/api-keys.

    sessionAuth:
      type: apiKey
      in: cookie
      name: wrenn_sid
      description: |
        Opaque session cookie set by POST /v1/auth/login, /v1/auth/activate, or
        the OAuth callback. HttpOnly, Secure, SameSite=Strict. Idle window 6h,
        absolute lifetime 24h. State-changing requests also require an
        X-CSRF-Token header matching the wrenn_csrf cookie (double-submit).
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit CSRF token whose value must match the wrenn_csrf cookie.
        Required on all non-GET requests authenticated via session cookie.
        Not required for API key auth.

    hostTokenAuth:
      type: apiKey
      in: header
      name: X-Host-Token
      description: Host JWT returned from POST /v1/hosts/register or POST /v1/hosts/auth/refresh. Valid for 7 days.
  schemas:
...
    CreateSnapshotRequest:
      type: object
      required: [sandbox_id]
      properties:
        sandbox_id:
          type: string
          description: ID of the running capsule to snapshot.
        name:
          type: string
          description: Name for the snapshot template. Auto-generated if omitted.
    Template:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [base, snapshot]
        vcpus:
          type: integer
          nullable: true
        memory_mb:
          type: integer
          nullable: true
        size_bytes:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        platform:
          type: boolean
          description: |
            True when the template is platform-managed (visible to all teams,
            e.g. the built-in `minimal` rootfs). False for team-owned
            snapshot templates.
        metadata:
          type: object
          additionalProperties: {type: string}
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...